Ones and Zeroes
In the computer world, use restricted resource you have to generate maximum benefit is what we always want to pursue.

For now, suppose you are a dominator of m 0s and n 1s respectively. On the other hand, there is an array with strings consisting of only 0s and 1s.

Now your task is to find the maximum number of strings that you can form with given m 0s and n 1s. Each 0 and 1 can be used at most once.

Note:
The given numbers of 0s and 1s will both not exceed 100
The size of given string array won't exceed 600.
Example 1:
Input: Array = {"10", "0001", "111001", "1", "0"}, m = 5, n = 3
Output: 4

Explanation: This are totally 4 strings can be formed by the using of 5 0s and 3 1s, which are “10,”0001”,”1”,”0”
Example 2:
Input: Array = {"10", "0", "1"}, m = 1, n = 1
Output: 2

Explanation: You could form "10", but then you'd have nothing left. Better form "0" and "1".

思路： 动态规划 dp[i][j]表示 i个0  和 j个1  能形成的最多字符串的数量。
           遍历所有字符串，先计算每个字符串字符0与1的个数。
          遍历dp时 ，从右下至左上，（若是从左上至右下，会导致重复迭代）。
          注：若某个字符串计算到dp[i][j]，i必须要>=0的个数；j必须要>=1的个数，才能更新dp，表示可能由dp[i-count[0]][j-count[1]]加1代替